mv in pix base

+++ include/motion.h
@@ -3,7 +3,7 @@
 
 #include "preesm.h"
 
-unsigned int SAD(unsigned char *ref, unsigned char *cur,int blockSize,int width);
+unsigned long SAD(unsigned char *ref, unsigned char *cur,int blockSize,int width);
 void fullSearch(int width, int height, int blockSize,IN int *blockIndex,IN unsigned char *input,IN unsigned char *reference,OUT int *output_x,OUT int *output_y);
